requirements.txt
numpy>=1.18
py-pde>=0.30
